+++ CHANGELOG.md
@@ -18,6 +18,9 @@ All notable changes to this project will be documented in this file.
 - Configuration for environment now pulls from the environment variable assigned in dev-tools configuration.
 - Scripts now live in a repository subfolder called './scripts/'.
 - Info Flags in newenv now use separate PowerShell functions with switch cases.
+- NewEnv environment can now be non-empty files.
+- NewEnv files are templated with {{ VARIABLE }} notation. Where variable is one of (YEAR, FIRST_NAME, LAST_NAME, CONTACT, ENV_NAME).
+
 
 ### Fixed
 - Added extra condition so newenv doesn't break before returning to initial directory on a failed run.