+++ env_templates/newenv/CHANGELOG.md
@@ -1,4 +1,4 @@
-<!-- / © Copyright YYYY NAME <CONTACT> -->
+<!-- / © Copyright {{ YEAR }} {{ FIRST_NAME }} {{ LAST_NAME }} <{{ CONTACT }}> -->
 # Changelog
 
 All notable changes to this project will be documented in this file.